Kleisli comprehensions
We talked about exploring the term calculus for comprehensions solely in the setting of a Cartesian monoidal category with a strong monad P. A good notion of a comprehension term here is syntactic morphisms in the "Kleisli category" of the monad. A morphism f : a -> b in Kl(P) is a morphism in the underlying category f : a -> P(b). I sketched out the comprehension calculus you can derive from the defining natural transformations of our categorical setting.

Relevant for #2